backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3